Update home.php
Changes have been made to the home page. I have embedded a webAPI for the tiktok videos using blockqoute, also changed the code for the embedded google maps with the use of iframe user would be able to click on direction and be redirected to google maps. CSS and bootstrap to make it stand out more. some contact information has been added below the google map.