Update ReviewsController.php
The change that was made was to the create function, which would be used to handle the Ajax request when fetching data to display on the home page. After I was able to fix the path in the routes for the review/create and test it works, I then created this function to take the data that the user inputs and then validate it. It will then push the data collected to the database and display the success message. Finally, it returns the user to the review page, and there is a button they click to go to the home.