Update home.php
I updated the Google API with the geolocation code, using the web API to implement the hardware API. Added the geolocation script with various functions that control the errors, location permission, and the latitude and longitude of users' position if location was granted. Also, the mapULR will display the user's location on google Maps.