Update home.php
The home page is fully dynamic, there is an embedded webapi and also a geolocation API that has been embedded with the Google API. I have used AJAX to dynamically fetch the reviews from the database. The Tiktok API is working well. This is the most recent updated code for the home page.